test: working on TCP server mock
