+++ test/doctests/TestSimpleTCPServer.cpp
@@ -1,16 +1,11 @@
-#include "CommonModes.h"
 #include "doctest.h"
-#include <stdexcept>
 
 #ifndef DOCTEST_CONFIG_NO_EXCEPTIONS
 
 DOCTEST_MAKE_STD_HEADERS_CLEAN_FROM_WARNINGS_ON_WALL_BEGIN
 #include "Manager.h"
-#include "SimpleTCPClient.h"
-#include "SimpleTCPServer.h"
-#include <iostream>
-#include <sstream>
-#include <thread>
+#include "SimpleTCPClientMock.h"
+#include "SimpleTCPServerMock.h"
 DOCTEST_MAKE_STD_HEADERS_CLEAN_FROM_WARNINGS_ON_WALL_END
 
 DOCTEST_MSVC_SUPPRESS_WARNING(4626)
@@ -24,8 +19,8 @@ DOCTEST_MSVC_SUPPRESS_WARNING(4626)
  * For doctest concurrency examples see
  * https://github.com/doctest/doctest/blob/master/examples/all_features/concurrency.cpp
  */
-TEST_CASE("Testing the SimpleTCPServer class" * doctest::timeout(10) *
-          doctest::skip(true)) {
+TEST_CASE("Testing the SimpleTCPServer class - unmocked" *
+          doctest::timeout(10) * doctest::skip(true)) {
   SimpleTCPServer server;
   CHECK(server.isStarted() == false);
 
@@ -81,106 +76,6 @@ TEST_CASE("Testing the SimpleTCPServer class" * doctest::timeout(10) *
   MESSAGE("[TEST] End of test");
 }
 
-class SimpleTCPServerMock : public SimpleTCPServer {
-public:
-  SimpleTCPServerMock() : SimpleTCPServer(){};
-  char buffer[32_K];
-  std::atomic<bool> clientConnection = false;
-  std::atomic<bool> clientIsConnected = false;
-
-  void start() override {
-    if (bool expected = false; !isStarted_.compare_exchange_strong(
-            expected, true)) { // thread safe comparison with exchange
-      return;
-    }
-    isStarted_ = true;
-
-    while (!stopSource_.stop_requested()) {
-      if (!clientConnection || clientIsConnected) {
-        // thread safe comparison, here clientIsConnected should be true
-        std::this_thread::sleep_for(
-            std::chrono::seconds(1)); // timer to avoid cpu burst
-        continue;
-      }
-
-      int client_socket = 1;
-      std::string client_ip = "localhost";
-      std::string client_info = "[" + client_ip + "] ";
-
-      SimpleLogger::LOG_INFO(client_info, " Client connection.");
-
-      clientHandlers_.emplace_back(
-          [this, client_socket, client_ip](std::stop_token st) {
-            handle_client(client_socket, client_ip, st);
-          },
-          stopSource_.get_token());
-
-      clientIsConnected = true;
-    }
-  }
-  void stop() override {
-    stopSource_.request_stop();
-    isStarted_ = false;
-  }
-  void handle_client(int client_socket, const std::string &client_ip,
-                     const std::stop_token &stoken) override {
-    std::string client_info = "[" + client_ip + "] ";
-    std::string lineBuffer;
-    while (!stoken.stop_requested()) {
-      memset(buffer, 0, 32_K);
-      std::this_thread::sleep_for(
-          std::chrono::seconds(2)); // time for client to write the buffer
-      size_t bytesReceived = strlen(buffer);
-      if (bytesReceived == 0) {
-        SimpleLogger::LOG_INFO(
-            client_info, SimpleLang::Message(Message::TCPClientDisconnected));
-
-        // Process any remaining data in lineBuffer
-        if (!lineBuffer.empty()) {
-          std::string line = lineBuffer;
-          lineBuffer.erase();
-          processLine(line, client_info);
-        }
-
-        return;
-      }
-
-      lineBuffer.append(buffer, bytesReceived);
-
-      processLineBuffer(lineBuffer, client_info);
-    }
-    SimpleLogger::LOG_INFO(client_info,
-                           SimpleLang::Message(Message::TCPClientDisconnected));
-    clientConnection = false;
-    clientIsConnected = false;
-  }
-};
-
-class SimpleTCPClientMock : public SimpleTCPClient {
-public:
-  SimpleTCPClientMock() : SimpleTCPClient(){};
-  SimpleTCPServerMock *server = nullptr;
-  void connect(const std::string &host, unsigned short port) override {}
-  void connect() {
-    if (server == nullptr) {
-      return;
-    }
-    server->clientConnection = true;
-  }
-  void send(const std::string &message) override {
-    if (server == nullptr) {
-      return;
-    }
-    std::strcpy(server->buffer, message.c_str());
-  }
-  void disconnect() override {
-    if (server == nullptr) {
-      return;
-    }
-    server->clientConnection = false;
-  }
-};
-
 TEST_CASE("Testing the SimpleTCPServer class - mocked" * doctest::timeout(10)) {
   auto server = new SimpleTCPServerMock();
   CHECK(server->isStarted() == false);
@@ -207,18 +102,24 @@ TEST_CASE("Testing the SimpleTCPServer class - mocked" * doctest::timeout(10)) {
   // Starting the client
   MESSAGE("[TEST] Starting the TCP client...");
   client.connect();
+  std::this_thread::sleep_for(
+      std::chrono::seconds(1)); // Allow server to connect
 
   // Message testing
   MESSAGE("[TEST] Sending first data...");
   client.send("0,0.08,0.43,0.90,0.42,1.00,0.62,0.33,0.38,0.10,0.07,0.00,0.00,"
               "0.38,0.00,0.00,1.00,0.92,0.00,1.00,0.00");
+  std::this_thread::sleep_for(std::chrono::seconds(1));
   MESSAGE("[TEST] Sending next data...");
   client.send("1,0.01,0.57,0.90,0.25,1.00,0.00,0.67,0.92,0.09,0.02,0.00,"
               "0.00,0.62,0.00,0.00,1.00,0.92,0.00,1.00,0.00");
+  std::this_thread::sleep_for(std::chrono::seconds(1));
   MESSAGE("[TEST] Sending next data...");
   client.send("oops");
+  std::this_thread::sleep_for(std::chrono::seconds(1));
   MESSAGE("[TEST] Sending next data...");
   client.send("           "); // trim check
+  std::this_thread::sleep_for(std::chrono::seconds(1));
   MESSAGE("[TEST] Sending next data...");
   client.send("   1.00,0.04,0.57,0.80,0.08,1.00,0.38,0.00,0.85,0.12,0.05,0.00,"
               "0.73,0.62,0.00,0.00,1.00,0.92,0.00,1.00,0.00      ");
